Fix Google Task Sync
When a time estimate is added in Omnifocus a time clause (`(for x minutes)`) is added to the task when it is pushed to google tasks. Presumably, this to the accomodate the GoogleTask <--> Reclaim.ai sync. However, this is added to the task title. This causes the titles not to match on subsequent sync, which was causing the task to be repeatedly pushed to Google as a new task.